feat: admin submitProducts 거절 기능 완료

+++ src/api/admin/admin.service.ts
@@ -1,6 +1,6 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { SubmitRepository } from '../submit/submit.repository';
-import { UpdateSubmitProducts } from './dto/admin.dto';
+import { RejectSubmitProducts, UpdateSubmitProducts } from './dto/admin.dto';
 import { ProductRepository } from '../product/product.repository';
 
 @Injectable()
@@ -28,20 +28,40 @@ export class AdminService {
 
   async allowSubmitProduct(updateSubmitProducts: UpdateSubmitProducts) {
     const { submitId, userUUID, shopId } = updateSubmitProducts;
+
     // 해당 유저가 제출한 submit status를 완료로 바꾸기 (status:1)
-    const userSubmitData = await this.submitRepository.findSubmitProductsBySubmitId(submitId);
+    const userSubmitData = await this.submitRepository.findSubmitProductsBySubmitId(submitId, userUUID);
+    if (!userSubmitData) throw new NotFoundException('Not Found User Submit Record');
     userSubmitData.status = 1; // success;
     await this.submitRepository.saveSubmit(userSubmitData);
+
     // shopId로 product_mapping 테이블을 서칭해서 현재 type:0 인 정보 delete하기
     const usingProductsMapping = await this.productRepository.findAllUsingProductMappingsByShopId(shopId);
+    if (usingProductsMapping.length === 0) throw new NotFoundException('Not Found Products Mapping');
     usingProductsMapping.map(async (data) => {
       await this.productRepository.deleteProducts(data.id);
     });
+
     // 유저의 uuid와 shopId로 product_mapping 테이블을 서칭해서 유저가 제안한 products의 type 0으로 업데이트하기
     const submitProductMappings = await this.productRepository.findAllSubmitProductMappingsByShopIdAndUUID(shopId, userUUID);
+    if (submitProductMappings.length === 0) throw new NotFoundException('Not Found Submit Products Mapping');
     submitProductMappings.map(async (data) => {
       await this.productRepository.updateToUsingProduct(data.id);
     });
+    return;
+  }
+
+  async rejectSubmitProduct(rejectSubmitProducts: RejectSubmitProducts) {
+    const { submitId, userUUID, shopId } = rejectSubmitProducts;
+
+    // 해당 유저가 제출한 submit status를 거절로 바꾸기 (status:2)
+    const userSubmitData = await this.submitRepository.findSubmitProductsBySubmitId(submitId, userUUID);
+    if (!userSubmitData) throw new NotFoundException('Not Found User Submit Record');
+    userSubmitData.status = 2; // success;
+    await this.submitRepository.saveSubmit(userSubmitData);
+
+    // shopId와 userUUID를 통해 유저가 제보한 판매목록 데이터 지우기
+    await this.productRepository.deleteAllSubmitProductMappingsByShopIdAndUUID(shopId, userUUID);
     return;
   }
